+++ test/crypto/formatting_test.dart
@@ -12,12 +12,4 @@ void main() {
     expect(hexToDartInt('0xff'), 0xff);
     expect(hexToDartInt('abcdef'), 0xabcdef);
   });
-
-  test('hexToBytes handles odd length', () {
-    expect(
-        hexToBytes(
-            '0xcf1daa12796907b13c6e104ee7185eb8b1e34db92b55a4655c147ff2fe5563c'),
-        hexToBytes(
-            '0x0cf1daa12796907b13c6e104ee7185eb8b1e34db92b55a4655c147ff2fe5563c'));
-  });
 }